# Conflicts:
#	src/Exceptionless.Core/Models/PersistentEvent.cs
#	src/Exceptionless.Core/Repositories/Configuration/Indexes/EventIndex.cs
#	src/Exceptionless.Core/Repositories/EventRepository.cs
#	src/Exceptionless.Core/Utility/EventStats.cs

+++ src/Exceptionless.Api/Controllers/Base/RepositoryApiController.cs
@@ -218,7 +218,13 @@ namespace Exceptionless.Api.Controllers {
         }
 
         protected virtual async Task<IEnumerable<string>> DeleteModelsAsync(ICollection<TModel> values) {
-            await _repository.RemoveAsync(values);
+            if (_supportsSoftDeletes) {
+                values.Cast<ISupportSoftDeletes>().ForEach(v => v.IsDeleted = true);
+                await _repository.SaveAsync(values);
+            } else {
+                await _repository.RemoveAsync(values);
+            }
+
             return new List<string>();
         }
     }
